Add LRGB tests
