+++ skfp/datasets/lrgb/__init__.py
@@ -1,0 +1,3 @@
+from .benchmark import load_lrgb_mol_benchmark, load_lrgb_mol_splits
+from .peptides_func import load_peptides_func
+from .peptides_struct import load_peptides_struct
